ImageClassification/requirements.txt
Flask==2.2.3
torch==2.1.0
torchvision==0.15.1
Pillow==9.5.0
requests==2.30.0
